--- db-1637120385451.json
@@ -1,100 +1,0 @@
-{
-  "products": [
-    {
-      "id": 0,
-      "name": "Café americano",
-      "price": 5,
-      "image": "https://raw.githubusercontent.com/DianaAraujoG/La-Magia-de-Mexico/main/img-Burger-Queen/cafe.png",
-      "type": "desayuno",
-      "dataEntry": "2021-11-16T12:25:43.511Z"
-    },
-    {
-      "id": 1,
-      "name": "Café con leche",
-      "price": 7,
-      "image": "https://raw.githubusercontent.com/DianaAraujoG/La-Magia-de-Mexico/main/img-Burger-Queen/cafe-con-leche.png",
-      "type": "desayuno",
-      "dataEntry": "2021-11-16T12:25:43.511Z"
-    },
-    {
-      "id": 2,
-      "name": "Sandwich de jamón y queso",
-      "price": 10,
-      "image": "https://raw.githubusercontent.com/DianaAraujoG/La-Magia-de-Mexico/main/img-Burger-Queen/sandwich.png",
-      "type": "desayuno",
-      "dataEntry": "2021-11-16T12:25:43.511Z"
-    },
-    {
-      "id": 3,
-      "name": "Jugo de frutas natural",
-      "price": 7,
-      "image": "https://raw.githubusercontent.com/DianaAraujoG/La-Magia-de-Mexico/main/img-Burger-Queen/orange-juice.png",
-      "type": "desayuno",
-      "dataEntry": "2021-11-16T12:25:43.511Z"
-    },
-    {
-      "id": 4,
-      "name": "Hamburguesa simple",
-      "price": 10,
-      "image": "https://raw.githubusercontent.com/DianaAraujoG/La-Magia-de-Mexico/main/img-Burger-Queen/hamburguesa.png",
-      "type": "almuerzo",
-      "dataEntry": "2021-11-16T12:25:43.511Z"
-    },
-    {
-      "id": 5,
-      "name": "Hamburguesa doble",
-      "price": 15,
-      "image": "https://raw.githubusercontent.com/DianaAraujoG/La-Magia-de-Mexico/main/img-Burger-Queen/hamburguesa%20(2).png",
-      "type": "almuerzo",
-      "dataEntry": "2021-11-16T12:25:43.511Z"
-    },
-    {
-      "id": 6,
-      "name": "Papas fritas",
-      "price": 5,
-      "image": "https://raw.githubusercontent.com/DianaAraujoG/La-Magia-de-Mexico/main/img-Burger-Queen/french-fries.png",
-      "type": "almuerzo",
-      "dataEntry": "2021-11-16T12:25:43.511Z"
-    },
-    {
-      "id": 7,
-      "name": "Aros de cebolla",
-      "price": 5,
-      "image": "https://raw.githubusercontent.com/DianaAraujoG/La-Magia-de-Mexico/main/img-Burger-Queen/aros-de-cebolla.png",
-      "type": "almuerzo",
-      "dataEntry": "2021-11-16T12:25:43.511Z"
-    },
-    {
-      "id": 8,
-      "name": "Agua 500ml",
-      "price": 5,
-      "image": "https://raw.githubusercontent.com/DianaAraujoG/La-Magia-de-Mexico/main/img-Burger-Queen/glass-of-water.png",
-      "type": "almuerzo",
-      "dataEntry": "2021-11-16T12:25:43.511Z"
-    },
-    {
-      "id": 9,
-      "name": "Agua 750ml",
-      "price": 7,
-      "image": "https://raw.githubusercontent.com/DianaAraujoG/La-Magia-de-Mexico/main/img-Burger-Queen/glass-of-water.png",
-      "type": "almuerzo",
-      "dataEntry": "2021-11-16T12:25:43.511Z"
-    },
-    {
-      "id": 10,
-      "name": "Bebida/Gaseosa 500ml",
-      "price": 5,
-      "image": "https://raw.githubusercontent.com/DianaAraujoG/La-Magia-de-Mexico/main/img-Burger-Queen/reajuste-salarial.png",
-      "type": "almuerzo",
-      "dataEntry": "2021-11-16T12:25:43.511Z"
-    },
-    {
-      "id": 11,
-      "name": "Bebida/Gaseosa 750ml",
-      "price": 7,
-      "image": "https://raw.githubusercontent.com/DianaAraujoG/La-Magia-de-Mexico/main/img-Burger-Queen/reajuste-salarial.png",
-      "type": "almuerzo",
-      "dataEntry": "2021-11-16T12:25:43.511Z"
-    }
-  ]
-}
