+++ src/components/HeaderWaitress.jsx
@@ -1,0 +1,35 @@
+import { useNavigate } from 'react-router-dom';
+import { auth } from '../FirebaseConfig';
+import logo from "../assets/Burger-Queen-logo.png";
+import { signOut } from "firebase/auth";
+import { Fragment } from 'react';
+import './style/Style.css';
+
+const HeaderWaitress = () => {
+    const navigate = useNavigate();
+    const logOut = async () => {
+        try {
+            await signOut(auth);
+            console.log('LogOut');
+            navigate("/");    
+        } catch (error) {
+            console.log(error.code);
+        }
+    }
+    return (
+        <Fragment>
+            <div className="headerW">
+                <div className= 'logoheader'><img src= {logo} alt= 'logo BurgerQueen'></img></div>
+                <button className='btnMenu'>
+                    <img src="https://raw.githubusercontent.com/DianaAraujoG/La-Magia-de-Mexico/main/img-Burger-Queen/menu_button.png" alt ='btnMenu'></img>
+                </button>
+                <button className='btnReady'>
+                    <img alt='btnReady' src="https://raw.githubusercontent.com/DianaAraujoG/La-Magia-de-Mexico/main/img-Burger-Queen/listo_button.png"></img>
+                </button>
+                <button className='btnlogOut' onClick={logOut}>Salir</button>
+            </div>
+        </Fragment>        
+    )
+}
+
+export default HeaderWaitress
